powerd/requirements.txt

llama-mqtt

